init: startup rework.
-We reworked that start up to accomplish 2 things.

--Booting with all four cores seems to cause instructions to
execute out of order. For example, I would calls into the
"Power Management Service" before it's initialization had been called
resulting in an initialization var mContext being null. Booting
with a single core removes that error altogether with absolutely no
change in the code. Strange??? but true. So on early-init we are
disabling all cores except cpu0. Cpu governor is also set to
"performance" during boot up.

--We moved all cpu tunables to after boot completed. At that point
we enable auto_hotplug which re-enables all cores.

NOTES: I have tested this serveral time and have gotten a consistant
boot time, from first appearance of the "Hisense" logo to the first
appearance of the lockscreen, of about 90 seconds. Approximately 15 - 17
seconds of that is the "Hisense" logo.